refactor(dwn-sdk-js): decompose ProtocolAuthorization into focused modules
Split the 1187-line ProtocolAuthorization class into 4 focused modules:

- protocol-authorization.ts (~400 lines): public orchestrators, fetchProtocolDefinition, getRuleSet
- protocol-authorization-action.ts (~330 lines): role verification, action matching, actor checks
- protocol-authorization-validation.ts (~390 lines): structural validation (type, path, tags, size)
- record-chain.ts (~100 lines): record chain construction and timestamp resolution

Also moved getTypeName utility to utils/protocols.ts alongside related helpers.

All 23 static methods preserved with identical signatures and behavior.
Backward compatible: no public API changes, only internal reorganization.
